Add missing Javadoc

+++ src/main/java/edu/ntnu/tobiasth/idatt2001/hospital/person/Person.java
@@ -23,7 +23,7 @@ public abstract class Person {
     this.lastName = lastName;
 
     try {
-      this.personalIdNumber = String.valueOf(Integer.parseInt(personalIdNumber));
+      this.personalIdNumber = String.valueOf(Long.parseLong(personalIdNumber));
     } catch (NumberFormatException e) {
       throw new IllegalArgumentException(String.format("%s is not a number.", personalIdNumber));
     }
